fixed merge bug - opponent laser properly targets player

+++ app/src/main/java/com/gamecodeschool/asteroidsfs/ObjectFactory.java
@@ -116,7 +116,7 @@ public class ObjectFactory {
         }
 
         //want to shoot in direction of player
-        public Laser getOpponentLaser(PointF oppPos, double playerAngle, int dmg) {
+        public Laser getOpponentLaser(PointF oppPos, float playerAngle, int dmg) {
                 SpaceObject temp = new SpaceObject(oppPos, playerAngle, defaultLaserVelocity,
                         screen.width / DIVISION_FACTOR / LASER_SIZE_FACTOR);
                 return new Laser(temp, dmg);
